requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.5
Django==3.2.1
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
phonenumbers==8.12.25
Pillow==8.2.0
pycparser==2.20
pytz==2021.1
six==1.16.0
sqlparse==0.4.1